sweetalert for delete button added

+++ server/public/scripts/client.js
@@ -99,27 +99,43 @@ function completeTask() {
 function deleteTask() {
     let taskId = $(this).data('id');
 
-    $.ajax({
-        method: 'DELETE',
-        url: '/tasks/' + taskId
-    }).then(function (){
-        getTasks();
-    }).catch(function (error) {
-        alert('error DELETEing task', error);
-    });
+    Swal.fire({
+        title: 'Are you sure you want to delete?',
+        showDenyButton: true,
+        confirmButtonText: 'Delete',
+    }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+            $.ajax({
+                method: 'DELETE',
+                url: '/tasks/' + taskId
+            }).then(function (){
+                getTasks();
+            }).catch(function (error) {
+                alert('error DELETEing task', error);
+            });
+        } else {
+            return;
+        }
+    })
+
+        
+
+    
 };
 
 function renderTable(tasks) {
+
     $('#taskListDiv').empty();
 
     for (let task of tasks) {
         $('#taskListDiv').append(`
             <div class="taskItem ${task.is_complete === true ? 'onHide' : ''}">
                 <h4>Owner: ${task.owner}</h4>
-                <section><span>Date Required Complete: </span>${task.f_date}</section>
-                <section><span>Task Status: </span>${task.is_complete == true ? 'Completed' : 'Not Complete'}</section>
-                <p><span>Details: </span>${task.details}</p>
-                <p>
+                <section><span class="lead">Date Required Complete: </span>${task.f_date}</section>
+                <section><span class="lead">Task Status: </span>${task.is_complete == true ? 'Completed' : 'Not Complete'}</section>
+                <section><span class="lead">Details: </span>${task.details}</section>
+                <p class="mt-3 mb-3">
                     <button class="btn btn-secondary completeButton" data-id="${task.id}" data-complete="${task.is_complete}">${task.is_complete === true ? 'Undo Complete' : 'Mark Completed'}</button>
                     <button class="btn btn-danger deleteButton" data-id="${task.id}">Delete</button>
                 </p>
